Finish two project setup
