+++ README.md
@@ -1,0 +1,7 @@
+# LEDSync
+ESP32-based music LED controller designed to synchronize LED lighting effects with music.
+
+# Authors
+- [Filip Kowalski](https://github.com/spookyless)
+- [Michał Broś](https://github.com/MBrosik)
+- [Krzysztof Wrona](https://github.com/rubikon02)
