--- client/main/modules/device_registerer/device_registerer.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#define expireTime 7 * 24 * 60 * 60 * 1000 // 7 days
-
-class DeviceRegisterer {
-public:
-   static void tryRegister();
-private:   
-   static void setClaim();
-};
