--- client/main/modules/gatt_server/gatt_server.cpp
@@ -1,967 +1,0 @@
-#include "driver/gpio.h"
-#include "esp_bt.h"
-#include "esp_bt_defs.h"
-#include "esp_bt_main.h"
-#include "esp_event.h"
-#include "esp_gap_ble_api.h"
-#include "esp_gatt_common_api.h"
-#include "esp_gatts_api.h"
-#include "esp_log.h"
-#include "esp_system.h"
-#include "esp_timer.h"
-#include "freertos/FreeRTOS.h"
-#include "freertos/event_groups.h"
-#include "freertos/queue.h"
-#include "freertos/task.h"
-#include "nvs_flash.h"
-#include "sdkconfig.h"
-#include <inttypes.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include "gatt_server.h"
-
-#define _OPEN_SYS_ITOA_EXT
-
-#define GATTS_TAG "LEDSYNC_SERVER"
-
-/// Declare the static function
-static void gatts_profile_a_event_handler(esp_gatts_cb_event_t event,
-                                          esp_gatt_if_t gatts_if,
-                                          esp_ble_gatts_cb_param_t *param);
-static void gatts_profile_b_event_handler(esp_gatts_cb_event_t event,
-                                          esp_gatt_if_t gatts_if,
-                                          esp_ble_gatts_cb_param_t *param);
-
-uint32_t state_counter = 0;
-
-uint8_t state_A = 0;
-#define GATTS_SERVICE_UUID_TEST_A 0xDEAD
-#define GATTS_CHAR_UUID_TEST_A 0xDE01
-#define GATTS_DESCR_UUID_TEST_A 0xDE11
-#define GATTS_NUM_HANDLE_TEST_A 4
-#define MQTT_TOPIC_A "ledsync/milestone02/A/state"
-
-uint8_t state_B = 0;
-#define GATTS_SERVICE_UUID_TEST_B 0xBEEF
-#define GATTS_CHAR_UUID_TEST_B 0xBE01
-#define GATTS_DESCR_UUID_TEST_B 0xBE11
-#define GATTS_NUM_HANDLE_TEST_B 4
-#define MQTT_TOPIC_B "ledsync/milestone02/B/state"
-
-#define TEST_DEVICE_NAME "LEDSYNC_SERVER"
-#define TEST_MANUFACTURER_DATA_LEN 17
-
-#define GATTS_DEMO_CHAR_VAL_LEN_MAX 0x40
-
-#define PREPARE_BUF_MAX_SIZE 1024
-
-static uint8_t char1_str[] = {0x11, 0x22, 0x33};
-static esp_gatt_char_prop_t a_property = 0;
-static esp_gatt_char_prop_t b_property = 0;
-
-static esp_attr_value_t gatts_demo_char1_val = {
-    .attr_max_len = GATTS_DEMO_CHAR_VAL_LEN_MAX,
-    .attr_len = sizeof(char1_str),
-    .attr_value = char1_str,
-};
-
-static uint8_t adv_config_done = 0;
-#define adv_config_flag (1 << 0)
-#define scan_rsp_config_flag (1 << 1)
-
-static uint8_t adv_service_uuid128[32] = {
-    /* LSB
-       <-------------------------------------------------------------------------------->
-       MSB */
-    // first uuid, 16bit, [12],[13] is the value
-    0xfb,
-    0x34,
-    0x9b,
-    0x5f,
-    0x80,
-    0x00,
-    0x00,
-    0x80,
-    0x00,
-    0x10,
-    0x00,
-    0x00,
-    0xAD,
-    0xDE,
-    0x00,
-    0x00,
-    // second uuid, 32bit, [12], [13], [14], [15] is the value
-    0xfb,
-    0x34,
-    0x9b,
-    0x5f,
-    0x80,
-    0x00,
-    0x00,
-    0x80,
-    0x00,
-    0x10,
-    0x00,
-    0x00,
-    0xEF,
-    0xBE,
-    0x00,
-    0x00,
-};
-
-// The length of adv data must be less than 31 bytes
-// static uint8_t test_manufacturer[TEST_MANUFACTURER_DATA_LEN] =  {0x12, 0x23,
-// 0x45, 0x56}; adv data
-static esp_ble_adv_data_t adv_data = {
-    .set_scan_rsp = false,
-    .include_name = true,
-    .include_txpower = false,
-    .min_interval = 0x0006, // slave connection min interval, Time =
-                            // min_interval * 1.25 msec
-    .max_interval = 0x0010, // slave connection max interval, Time =
-                            // max_interval * 1.25 msec
-    .appearance = 0x00,
-    .manufacturer_len = 0,       // TEST_MANUFACTURER_DATA_LEN,
-    .p_manufacturer_data = NULL, //&test_manufacturer[0],
-    .service_data_len = 0,
-    .p_service_data = NULL,
-    .service_uuid_len = sizeof(adv_service_uuid128),
-    .p_service_uuid = adv_service_uuid128,
-    .flag = (ESP_BLE_ADV_FLAG_GEN_DISC | ESP_BLE_ADV_FLAG_BREDR_NOT_SPT),
-};
-// scan response data
-static esp_ble_adv_data_t scan_rsp_data = {
-    .set_scan_rsp = true,
-    .include_name = true,
-    .include_txpower = true,
-    //.min_interval = 0x0006,
-    //.max_interval = 0x0010,
-    .appearance = 0x00,
-    .manufacturer_len = 0,       // TEST_MANUFACTURER_DATA_LEN,
-    .p_manufacturer_data = NULL, //&test_manufacturer[0],
-    .service_data_len = 0,
-    .p_service_data = NULL,
-    .service_uuid_len = sizeof(adv_service_uuid128),
-    .p_service_uuid = adv_service_uuid128,
-    .flag = (ESP_BLE_ADV_FLAG_GEN_DISC | ESP_BLE_ADV_FLAG_BREDR_NOT_SPT),
-};
-
-static esp_ble_adv_params_t adv_params = {
-    .adv_int_min = 0x20,
-    .adv_int_max = 0x40,
-    .adv_type = ADV_TYPE_IND,
-    .own_addr_type = BLE_ADDR_TYPE_PUBLIC,
-    //.peer_addr            =
-    //.peer_addr_type       =
-    .channel_map = ADV_CHNL_ALL,
-    .adv_filter_policy = ADV_FILTER_ALLOW_SCAN_ANY_CON_ANY,
-};
-
-#define PROFILE_NUM 2
-#define PROFILE_A_APP_ID 0
-#define PROFILE_B_APP_ID 1
-
-struct gatts_profile_inst {
-  esp_gatts_cb_t gatts_cb;
-  uint16_t gatts_if;
-  uint16_t app_id;
-  uint16_t conn_id;
-  uint16_t service_handle;
-  esp_gatt_srvc_id_t service_id;
-  uint16_t char_handle;
-  esp_bt_uuid_t char_uuid;
-  esp_gatt_perm_t perm;
-  esp_gatt_char_prop_t property;
-  uint16_t descr_handle;
-  esp_bt_uuid_t descr_uuid;
-};
-
-/* One gatt-based profile one app_id and one gatts_if, this array will store the
- * gatts_if returned by ESP_GATTS_REG_EVT */
-static struct gatts_profile_inst gl_profile_tab[PROFILE_NUM] = {
-    [PROFILE_A_APP_ID] =
-        {
-            .gatts_cb = gatts_profile_a_event_handler,
-            .gatts_if = ESP_GATT_IF_NONE, /* Not get the gatt_if, so initial is
-                                             ESP_GATT_IF_NONE */
-        },
-    [PROFILE_B_APP_ID] =
-        {
-            .gatts_cb =
-                gatts_profile_b_event_handler, /* This demo does not implement,
-                                                  similar as profile A */
-            .gatts_if = ESP_GATT_IF_NONE, /* Not get the gatt_if, so initial is
-                                             ESP_GATT_IF_NONE */
-        },
-};
-
-typedef struct {
-  uint8_t *prepare_buf;
-  int prepare_len;
-} prepare_type_env_t;
-
-static prepare_type_env_t a_prepare_write_env;
-static prepare_type_env_t b_prepare_write_env;
-
-void example_write_event_env(esp_gatt_if_t gatts_if,
-                             prepare_type_env_t *prepare_write_env,
-                             esp_ble_gatts_cb_param_t *param);
-void example_exec_write_event_env(prepare_type_env_t *prepare_write_env,
-                                  esp_ble_gatts_cb_param_t *param);
-
-static void gap_event_handler(esp_gap_ble_cb_event_t event,
-                              esp_ble_gap_cb_param_t *param) {
-  switch (event) {
-    // When advertising data set complete, the event comes
-    case ESP_GAP_BLE_ADV_DATA_SET_COMPLETE_EVT: {
-        adv_config_done &= (~adv_config_flag);
-        if (adv_config_done == 0) {
-            esp_ble_gap_start_advertising(&adv_params);
-        }
-        break;
-    }
-
-    // When scan response data set complete, the event comes
-    case ESP_GAP_BLE_SCAN_RSP_DATA_SET_COMPLETE_EVT: {
-        adv_config_done &= (~scan_rsp_config_flag);
-        if (adv_config_done == 0) {
-        esp_ble_gap_start_advertising(&adv_params);
-        }
-        break;
-    }
-
-    // When start advertising complete, the event comes
-    case ESP_GAP_BLE_ADV_START_COMPLETE_EVT: {
-        // advertising start complete event to indicate advertising start
-        // successfully or failed
-        if (param->adv_start_cmpl.status != ESP_BT_STATUS_SUCCESS) {
-        ESP_LOGE(GATTS_TAG, "Advertising start failed\n");
-        }
-        break;
-    }
-
-    // When stop adv complete, the event comes
-    case ESP_GAP_BLE_ADV_STOP_COMPLETE_EVT: {
-        if (param->adv_stop_cmpl.status != ESP_BT_STATUS_SUCCESS) {
-        ESP_LOGE(GATTS_TAG, "Advertising stop failed\n");
-        } else {
-        ESP_LOGI(GATTS_TAG, "Stop adv successfully\n");
-        }
-        break;
-    }
-
-    // When update connection parameters complete, the event comes
-    case ESP_GAP_BLE_UPDATE_CONN_PARAMS_EVT: {
-        ESP_LOGI(
-            GATTS_TAG,
-            "update connection params status = %d, conn_int = %d, latency = %d, "
-            "timeout = %d",
-            param->update_conn_params.status, param->update_conn_params.conn_int,
-            param->update_conn_params.latency, param->update_conn_params.timeout);
-        break;
-    }
-
-    // When set packet length complete, the event comes
-    case ESP_GAP_BLE_SET_PKT_LENGTH_COMPLETE_EVT: {
-        ESP_LOGI(GATTS_TAG, "packet length updated: rx = %d, tx = %d, status = %d",
-                param->pkt_data_length_cmpl.params.rx_len,
-                param->pkt_data_length_cmpl.params.tx_len,
-                param->pkt_data_length_cmpl.status);
-        break;
-    }
-    default: {
-        break;
-    }
-  }
-}
-
-void example_write_event_env(esp_gatt_if_t gatts_if,
-                             prepare_type_env_t *prepare_write_env,
-                             esp_ble_gatts_cb_param_t *param) {
-  esp_gatt_status_t status = ESP_GATT_OK;
-  if (param->write.need_rsp) {
-    if (param->write.is_prep) {
-      if (param->write.offset > PREPARE_BUF_MAX_SIZE) {
-        status = ESP_GATT_INVALID_OFFSET;
-      } else if ((param->write.offset + param->write.len) >
-                 PREPARE_BUF_MAX_SIZE) {
-        status = ESP_GATT_INVALID_ATTR_LEN;
-      }
-      if (status == ESP_GATT_OK && prepare_write_env->prepare_buf == NULL) {
-        prepare_write_env->prepare_buf =
-            (uint8_t *)malloc(PREPARE_BUF_MAX_SIZE * sizeof(uint8_t));
-        prepare_write_env->prepare_len = 0;
-        if (prepare_write_env->prepare_buf == NULL) {
-          ESP_LOGE(GATTS_TAG, "Gatt_server prep no mem");
-          status = ESP_GATT_NO_RESOURCES;
-        }
-      }
-
-      esp_gatt_rsp_t *gatt_rsp =
-          (esp_gatt_rsp_t *)malloc(sizeof(esp_gatt_rsp_t));
-      if (gatt_rsp) {
-        gatt_rsp->attr_value.len = param->write.len;
-        gatt_rsp->attr_value.handle = param->write.handle;
-        gatt_rsp->attr_value.offset = param->write.offset;
-        gatt_rsp->attr_value.auth_req = ESP_GATT_AUTH_REQ_NONE;
-        memcpy(gatt_rsp->attr_value.value, param->write.value,
-               param->write.len);
-        esp_err_t response_err = esp_ble_gatts_send_response(
-            gatts_if, param->write.conn_id, param->write.trans_id, status,
-            gatt_rsp);
-        if (response_err != ESP_OK) {
-          ESP_LOGE(GATTS_TAG, "Send response error\n");
-        }
-        free(gatt_rsp);
-      } else {
-        ESP_LOGE(GATTS_TAG,
-                 "malloc failed, no resource to send response error\n");
-        status = ESP_GATT_NO_RESOURCES;
-      }
-      if (status != ESP_GATT_OK) {
-        return;
-      }
-      memcpy(prepare_write_env->prepare_buf + param->write.offset,
-             param->write.value, param->write.len);
-      prepare_write_env->prepare_len += param->write.len;
-
-    } else {
-      esp_ble_gatts_send_response(gatts_if, param->write.conn_id,
-                                  param->write.trans_id, status, NULL);
-    }
-  }
-}
-
-void example_exec_write_event_env(prepare_type_env_t *prepare_write_env,
-                                  esp_ble_gatts_cb_param_t *param) {
-  if (param->exec_write.exec_write_flag == ESP_GATT_PREP_WRITE_EXEC) {
-    ESP_LOG_BUFFER_HEX(GATTS_TAG, prepare_write_env->prepare_buf,
-                       prepare_write_env->prepare_len);
-  } else {
-    ESP_LOGI(GATTS_TAG, "Prepare write cancel");
-  }
-  if (prepare_write_env->prepare_buf) {
-    free(prepare_write_env->prepare_buf);
-    prepare_write_env->prepare_buf = NULL;
-  }
-  prepare_write_env->prepare_len = 0;
-}
-
-static void gatts_profile_a_event_handler(esp_gatts_cb_event_t event,
-                                          esp_gatt_if_t gatts_if,
-                                          esp_ble_gatts_cb_param_t *param) {
-  switch (event) {
-    case ESP_GATTS_REG_EVT: {
-        ESP_LOGI(GATTS_TAG, "REGISTER_APP_EVT, status %d, app_id %d\n",
-                param->reg.status, param->reg.app_id);
-        gl_profile_tab[PROFILE_A_APP_ID].service_id.is_primary = true;
-        gl_profile_tab[PROFILE_A_APP_ID].service_id.id.inst_id = 0x00;
-        gl_profile_tab[PROFILE_A_APP_ID].service_id.id.uuid.len = ESP_UUID_LEN_16;
-        gl_profile_tab[PROFILE_A_APP_ID].service_id.id.uuid.uuid.uuid16 =
-            GATTS_SERVICE_UUID_TEST_A;
-
-        esp_err_t set_dev_name_ret = esp_ble_gap_set_device_name(TEST_DEVICE_NAME);
-        if (set_dev_name_ret) {
-        ESP_LOGE(GATTS_TAG, "set device name failed, error code = %x",
-                set_dev_name_ret);
-        }
-
-        // config adv data
-        esp_err_t ret = esp_ble_gap_config_adv_data(&adv_data);
-        if (ret) {
-        ESP_LOGE(GATTS_TAG, "config adv data failed, error code = %x", ret);
-        }
-        adv_config_done |= adv_config_flag;
-        // config scan response data
-        ret = esp_ble_gap_config_adv_data(&scan_rsp_data);
-        if (ret) {
-        ESP_LOGE(GATTS_TAG, "config scan response data failed, error code = %x",
-                ret);
-        }
-        adv_config_done |= scan_rsp_config_flag;
-
-        esp_ble_gatts_create_service(gatts_if,
-                                    &gl_profile_tab[PROFILE_A_APP_ID].service_id,
-                                    GATTS_NUM_HANDLE_TEST_A);
-        break;
-    }
-    case ESP_GATTS_READ_EVT: {
-        ESP_LOGI(GATTS_TAG,
-                "GATT_READ_EVT, conn_id %d, trans_id %" PRIu32 ", handle %d\n",
-                param->read.conn_id, param->read.trans_id, param->read.handle);
-        esp_gatt_rsp_t rsp;
-        memset(&rsp, 0, sizeof(esp_gatt_rsp_t));
-        rsp.attr_value.handle = param->read.handle;
-        rsp.attr_value.len = 1;
-        rsp.attr_value.value[0] = state_A;
-        esp_ble_gatts_send_response(gatts_if, param->read.conn_id,
-                                    param->read.trans_id, ESP_GATT_OK, &rsp);
-        break;
-    }
-    case ESP_GATTS_WRITE_EVT: {
-        ESP_LOGI(GATTS_TAG,
-                "Characteristic write, conn_id %d, trans_id %" PRIu32
-                ", handle %d",
-                param->write.conn_id, param->write.trans_id, param->write.handle);
-        if (!param->write.is_prep) {
-        ESP_LOGI(GATTS_TAG, "value len %d, value ", param->write.len);
-        ESP_LOG_BUFFER_HEX(GATTS_TAG, param->write.value, param->write.len);
-        if (gl_profile_tab[PROFILE_A_APP_ID].descr_handle ==
-                param->write.handle &&
-            param->write.len == 2) {
-            uint16_t descr_value =
-                param->write.value[1] << 8 | param->write.value[0];
-            if (descr_value == 0x0001) {
-            if (a_property & ESP_GATT_CHAR_PROP_BIT_NOTIFY) {
-                ESP_LOGI(GATTS_TAG, "Notification enable");
-                uint8_t notify_data[15];
-                for (int i = 0; i < sizeof(notify_data); ++i) {
-                notify_data[i] = i % 0xff;
-                }
-                // the size of notify_data[] need less than MTU size
-                esp_ble_gatts_send_indicate(
-                    gatts_if, param->write.conn_id,
-                    gl_profile_tab[PROFILE_A_APP_ID].char_handle,
-                    sizeof(notify_data), notify_data, false);
-            }
-            } else if (descr_value == 0x0002) {
-            if (a_property & ESP_GATT_CHAR_PROP_BIT_INDICATE) {
-                ESP_LOGI(GATTS_TAG, "Indication enable");
-                uint8_t indicate_data[15];
-                for (int i = 0; i < sizeof(indicate_data); ++i) {
-                indicate_data[i] = i % 0xff;
-                }
-                // the size of indicate_data[] need less than MTU size
-                esp_ble_gatts_send_indicate(
-                    gatts_if, param->write.conn_id,
-                    gl_profile_tab[PROFILE_A_APP_ID].char_handle,
-                    sizeof(indicate_data), indicate_data, true);
-            }
-            } else if (descr_value == 0x0000) {
-            ESP_LOGI(GATTS_TAG, "Notification/Indication disable");
-            } else {
-            ESP_LOGE(GATTS_TAG, "Unknown descriptor value");
-            ESP_LOG_BUFFER_HEX(GATTS_TAG, param->write.value, param->write.len);
-            }
-        }
-        }
-        example_write_event_env(gatts_if, &a_prepare_write_env, param);
-        break;
-    }
-    case ESP_GATTS_EXEC_WRITE_EVT: {
-        ESP_LOGI(GATTS_TAG, "Execute write");
-        esp_ble_gatts_send_response(gatts_if, param->write.conn_id,
-                                    param->write.trans_id, ESP_GATT_OK, NULL);
-        example_exec_write_event_env(&a_prepare_write_env, param);
-        break;
-    }
-    case ESP_GATTS_MTU_EVT: {
-        ESP_LOGI(GATTS_TAG, "ESP_GATTS_MTU_EVT, MTU %d", param->mtu.mtu);
-        break;
-    }
-    case ESP_GATTS_UNREG_EVT: {
-        break;
-    }
-    case ESP_GATTS_CREATE_EVT: {
-        ESP_LOGI(GATTS_TAG, "CREATE_SERVICE_EVT, status %d,  service_handle %d\n",
-                param->create.status, param->create.service_handle);
-        gl_profile_tab[PROFILE_A_APP_ID].service_handle =
-            param->create.service_handle;
-        gl_profile_tab[PROFILE_A_APP_ID].char_uuid.len = ESP_UUID_LEN_16;
-        gl_profile_tab[PROFILE_A_APP_ID].char_uuid.uuid.uuid16 =
-            GATTS_CHAR_UUID_TEST_A;
-
-        esp_ble_gatts_start_service(
-            gl_profile_tab[PROFILE_A_APP_ID].service_handle);
-        a_property = ESP_GATT_CHAR_PROP_BIT_READ | ESP_GATT_CHAR_PROP_BIT_NOTIFY;
-        esp_err_t add_char_ret =
-            esp_ble_gatts_add_char(gl_profile_tab[PROFILE_A_APP_ID].service_handle,
-                                &gl_profile_tab[PROFILE_A_APP_ID].char_uuid,
-                                ESP_GATT_PERM_READ | ESP_GATT_PERM_WRITE,
-                                a_property, &gatts_demo_char1_val, NULL);
-        if (add_char_ret) {
-        ESP_LOGE(GATTS_TAG, "add char failed, error code =%x", add_char_ret);
-        }
-        break;
-    }
-    case ESP_GATTS_ADD_INCL_SRVC_EVT: {
-        break;
-    }
-    case ESP_GATTS_ADD_CHAR_EVT: {
-        uint16_t length = 0;
-        const uint8_t *prf_char;
-
-        ESP_LOGI(GATTS_TAG,
-                "ADD_CHAR_EVT, status %d,  attr_handle %d, service_handle %d\n",
-                param->add_char.status, param->add_char.attr_handle,
-                param->add_char.service_handle);
-        gl_profile_tab[PROFILE_A_APP_ID].char_handle = param->add_char.attr_handle;
-        gl_profile_tab[PROFILE_A_APP_ID].descr_uuid.len = ESP_UUID_LEN_16;
-        gl_profile_tab[PROFILE_A_APP_ID].descr_uuid.uuid.uuid16 =
-            ESP_GATT_UUID_CHAR_CLIENT_CONFIG;
-        esp_err_t get_attr_ret = esp_ble_gatts_get_attr_value(
-            param->add_char.attr_handle, &length, &prf_char);
-        if (get_attr_ret == ESP_FAIL) {
-        ESP_LOGE(GATTS_TAG, "ILLEGAL HANDLE");
-        }
-
-        ESP_LOGI(GATTS_TAG, "the gatts demo char length = %x\n", length);
-        for (int i = 0; i < length; i++) {
-        ESP_LOGI(GATTS_TAG, "prf_char[%x] =%x\n", i, prf_char[i]);
-        }
-        esp_err_t add_descr_ret = esp_ble_gatts_add_char_descr(
-            gl_profile_tab[PROFILE_A_APP_ID].service_handle,
-            &gl_profile_tab[PROFILE_A_APP_ID].descr_uuid,
-            ESP_GATT_PERM_READ | ESP_GATT_PERM_WRITE, NULL, NULL);
-        if (add_descr_ret) {
-        ESP_LOGE(GATTS_TAG, "add char descr failed, error code =%x",
-                add_descr_ret);
-        }
-        break;
-    }
-    case ESP_GATTS_ADD_CHAR_DESCR_EVT: {
-        gl_profile_tab[PROFILE_A_APP_ID].descr_handle =
-            param->add_char_descr.attr_handle;
-        ESP_LOGI(GATTS_TAG,
-                "ADD_DESCR_EVT, status %d, attr_handle %d, service_handle %d\n",
-                param->add_char_descr.status, param->add_char_descr.attr_handle,
-                param->add_char_descr.service_handle);
-        break;
-    }
-    case ESP_GATTS_DELETE_EVT: {
-        break;
-    }
-    case ESP_GATTS_START_EVT: {
-        ESP_LOGI(GATTS_TAG, "SERVICE_START_EVT, status %d, service_handle %d\n",
-                param->start.status, param->start.service_handle);
-        break;
-    }
-    case ESP_GATTS_STOP_EVT: {
-        break;
-    }
-    case ESP_GATTS_CONNECT_EVT: {
-        esp_ble_conn_update_params_t conn_params = {0};
-        memcpy(conn_params.bda, param->connect.remote_bda, sizeof(esp_bd_addr_t));
-        /* For the IOS system, please reference the apple official documents about
-        * the ble connection parameters restrictions. */
-        conn_params.latency = 0;
-        conn_params.max_int = 0x20; // max_int = 0x20*1.25ms = 40ms
-        conn_params.min_int = 0x10; // min_int = 0x10*1.25ms = 20ms
-        conn_params.timeout = 400;  // timeout = 400*10ms = 4000ms
-        ESP_LOGI(GATTS_TAG,
-                "ESP_GATTS_CONNECT_EVT, conn_id %d, remote "
-                "%02x:%02x:%02x:%02x:%02x:%02x:",
-                param->connect.conn_id, param->connect.remote_bda[0],
-                param->connect.remote_bda[1], param->connect.remote_bda[2],
-                param->connect.remote_bda[3], param->connect.remote_bda[4],
-                param->connect.remote_bda[5]);
-        gl_profile_tab[PROFILE_A_APP_ID].conn_id = param->connect.conn_id;
-        // start sent the update connection parameters to the peer device.
-        esp_ble_gap_update_conn_params(&conn_params);
-        break;
-    }
-    case ESP_GATTS_DISCONNECT_EVT: {
-        ESP_LOGI(GATTS_TAG, "ESP_GATTS_DISCONNECT_EVT, disconnect reason 0x%x",
-                param->disconnect.reason);
-        esp_ble_gap_start_advertising(&adv_params);
-        break;
-    }
-    case ESP_GATTS_CONF_EVT: {
-        ESP_LOGI(GATTS_TAG, "ESP_GATTS_CONF_EVT, status %d attr_handle %d",
-                param->conf.status, param->conf.handle);
-        if (param->conf.status != ESP_GATT_OK) {
-        esp_log_buffer_hex(GATTS_TAG, param->conf.value, param->conf.len);
-        }
-        break;
-    }
-    default: {
-        break;
-    }
-  }
-}
-
-static void gatts_profile_b_event_handler(esp_gatts_cb_event_t event,
-                                          esp_gatt_if_t gatts_if,
-                                          esp_ble_gatts_cb_param_t *param) {
-  switch (event) {
-    case ESP_GATTS_REG_EVT: {
-        ESP_LOGI(GATTS_TAG, "REGISTER_APP_EVT, status %d, app_id %d\n",
-                param->reg.status, param->reg.app_id);
-        gl_profile_tab[PROFILE_B_APP_ID].service_id.is_primary = true;
-        gl_profile_tab[PROFILE_B_APP_ID].service_id.id.inst_id = 0x00;
-        gl_profile_tab[PROFILE_B_APP_ID].service_id.id.uuid.len = ESP_UUID_LEN_16;
-        gl_profile_tab[PROFILE_B_APP_ID].service_id.id.uuid.uuid.uuid16 =
-            GATTS_SERVICE_UUID_TEST_B;
-
-        esp_ble_gatts_create_service(gatts_if,
-                                    &gl_profile_tab[PROFILE_B_APP_ID].service_id,
-                                    GATTS_NUM_HANDLE_TEST_B);
-        break;
-    }
-    case ESP_GATTS_READ_EVT: {
-        ESP_LOGI(GATTS_TAG,
-                "GATT_READ_EVT, conn_id %d, trans_id %" PRIu32 ", handle %d\n",
-                param->read.conn_id, param->read.trans_id, param->read.handle);
-        esp_gatt_rsp_t rsp;
-        memset(&rsp, 0, sizeof(esp_gatt_rsp_t));
-        rsp.attr_value.handle = param->read.handle;
-        rsp.attr_value.len = 1;
-        rsp.attr_value.value[0] = state_B;
-        esp_ble_gatts_send_response(gatts_if, param->read.conn_id,
-                                    param->read.trans_id, ESP_GATT_OK, &rsp);
-        break;
-    }
-    case ESP_GATTS_WRITE_EVT: {
-        ESP_LOGI(GATTS_TAG,
-                "Characteristic write, conn_id %d, trans_id %" PRIu32
-                ", handle %d",
-                param->write.conn_id, param->write.trans_id, param->write.handle);
-        if (!param->write.is_prep) {
-        ESP_LOGI(GATTS_TAG, "value len %d, value ", param->write.len);
-        ESP_LOG_BUFFER_HEX(GATTS_TAG, param->write.value, param->write.len);
-        if (gl_profile_tab[PROFILE_B_APP_ID].descr_handle ==
-                param->write.handle &&
-            param->write.len == 2) {
-            uint16_t descr_value =
-                param->write.value[1] << 8 | param->write.value[0];
-            if (descr_value == 0x0001) {
-            if (b_property & ESP_GATT_CHAR_PROP_BIT_NOTIFY) {
-                ESP_LOGI(GATTS_TAG, "Notification enable");
-                uint8_t notify_data[15];
-                for (int i = 0; i < sizeof(notify_data); ++i) {
-                notify_data[i] = i % 0xff;
-                }
-                // the size of notify_data[] need less than MTU size
-                esp_ble_gatts_send_indicate(
-                    gatts_if, param->write.conn_id,
-                    gl_profile_tab[PROFILE_B_APP_ID].char_handle,
-                    sizeof(notify_data), notify_data, false);
-            }
-            } else if (descr_value == 0x0002) {
-            if (b_property & ESP_GATT_CHAR_PROP_BIT_INDICATE) {
-                ESP_LOGI(GATTS_TAG, "Indication enable");
-                uint8_t indicate_data[15];
-                for (int i = 0; i < sizeof(indicate_data); ++i) {
-                indicate_data[i] = i % 0xff;
-                }
-                // the size of indicate_data[] need less than MTU size
-                esp_ble_gatts_send_indicate(
-                    gatts_if, param->write.conn_id,
-                    gl_profile_tab[PROFILE_B_APP_ID].char_handle,
-                    sizeof(indicate_data), indicate_data, true);
-            }
-            } else if (descr_value == 0x0000) {
-            ESP_LOGI(GATTS_TAG, "Notification/Indication disable");
-            } else {
-            ESP_LOGE(GATTS_TAG, "Unknown value");
-            }
-        }
-        }
-        example_write_event_env(gatts_if, &b_prepare_write_env, param);
-        break;
-    }
-    case ESP_GATTS_EXEC_WRITE_EVT: {
-        ESP_LOGI(GATTS_TAG, "Execute write");
-        esp_ble_gatts_send_response(gatts_if, param->write.conn_id,
-                                    param->write.trans_id, ESP_GATT_OK, NULL);
-        example_exec_write_event_env(&b_prepare_write_env, param);
-        break;
-    }
-    case ESP_GATTS_MTU_EVT: {
-        ESP_LOGI(GATTS_TAG, "ESP_GATTS_MTU_EVT, MTU %d", param->mtu.mtu);
-        break;
-    }
-    case ESP_GATTS_UNREG_EVT: {
-        break;
-    }
-    case ESP_GATTS_CREATE_EVT: {
-        ESP_LOGI(GATTS_TAG, "CREATE_SERVICE_EVT, status %d,  service_handle %d\n",
-                param->create.status, param->create.service_handle);
-        gl_profile_tab[PROFILE_B_APP_ID].service_handle =
-            param->create.service_handle;
-        gl_profile_tab[PROFILE_B_APP_ID].char_uuid.len = ESP_UUID_LEN_16;
-        gl_profile_tab[PROFILE_B_APP_ID].char_uuid.uuid.uuid16 =
-            GATTS_CHAR_UUID_TEST_B;
-
-        esp_ble_gatts_start_service(
-            gl_profile_tab[PROFILE_B_APP_ID].service_handle);
-        b_property = ESP_GATT_CHAR_PROP_BIT_READ | ESP_GATT_CHAR_PROP_BIT_NOTIFY;
-        esp_err_t add_char_ret = esp_ble_gatts_add_char(
-            gl_profile_tab[PROFILE_B_APP_ID].service_handle,
-            &gl_profile_tab[PROFILE_B_APP_ID].char_uuid,
-            ESP_GATT_PERM_READ | ESP_GATT_PERM_WRITE, b_property, NULL, NULL);
-        if (add_char_ret) {
-        ESP_LOGE(GATTS_TAG, "add char failed, error code =%x", add_char_ret);
-        }
-        break;
-    }
-    case ESP_GATTS_ADD_INCL_SRVC_EVT: {
-        break;
-    }
-    case ESP_GATTS_ADD_CHAR_EVT: {
-        ESP_LOGI(GATTS_TAG,
-                "ADD_CHAR_EVT, status %d,  attr_handle %d, service_handle %d\n",
-                param->add_char.status, param->add_char.attr_handle,
-                param->add_char.service_handle);
-
-        gl_profile_tab[PROFILE_B_APP_ID].char_handle = param->add_char.attr_handle;
-        gl_profile_tab[PROFILE_B_APP_ID].descr_uuid.len = ESP_UUID_LEN_16;
-        gl_profile_tab[PROFILE_B_APP_ID].descr_uuid.uuid.uuid16 =
-            ESP_GATT_UUID_CHAR_CLIENT_CONFIG;
-        esp_ble_gatts_add_char_descr(
-            gl_profile_tab[PROFILE_B_APP_ID].service_handle,
-            &gl_profile_tab[PROFILE_B_APP_ID].descr_uuid,
-            ESP_GATT_PERM_READ | ESP_GATT_PERM_WRITE, NULL, NULL);
-        break;
-    }
-    case ESP_GATTS_ADD_CHAR_DESCR_EVT: {
-        gl_profile_tab[PROFILE_B_APP_ID].descr_handle =
-            param->add_char_descr.attr_handle;
-        ESP_LOGI(GATTS_TAG,
-                "ADD_DESCR_EVT, status %d, attr_handle %d, service_handle %d\n",
-                param->add_char_descr.status, param->add_char_descr.attr_handle,
-                param->add_char_descr.service_handle);
-        break;
-    }
-    case ESP_GATTS_DELETE_EVT: {
-        break;
-    }
-    case ESP_GATTS_START_EVT: {
-        ESP_LOGI(GATTS_TAG, "SERVICE_START_EVT, status %d, service_handle %d\n",
-                param->start.status, param->start.service_handle);
-        break;
-    }
-    case ESP_GATTS_STOP_EVT: {
-        break;
-    }
-    case ESP_GATTS_CONNECT_EVT: {
-        ESP_LOGI(GATTS_TAG,
-                "CONNECT_EVT, conn_id %d, remote %02x:%02x:%02x:%02x:%02x:%02x:",
-                param->connect.conn_id, param->connect.remote_bda[0],
-                param->connect.remote_bda[1], param->connect.remote_bda[2],
-                param->connect.remote_bda[3], param->connect.remote_bda[4],
-                param->connect.remote_bda[5]);
-        gl_profile_tab[PROFILE_B_APP_ID].conn_id = param->connect.conn_id;
-        break;
-    }
-    case ESP_GATTS_CONF_EVT: {
-        ESP_LOGI(GATTS_TAG, "ESP_GATTS_CONF_EVT status %d attr_handle %d",
-                param->conf.status, param->conf.handle);
-        if (param->conf.status != ESP_GATT_OK) {
-        esp_log_buffer_hex(GATTS_TAG, param->conf.value, param->conf.len);
-        }
-        break;
-    }
-    default: {
-        break;
-    }
-  }
-}
-
-void gatts_event_handler(esp_gatts_cb_event_t event,
-                                esp_gatt_if_t gatts_if,
-                                esp_ble_gatts_cb_param_t *param) {
-  /* If event is register event, store the gatts_if for each profile */
-  if (event == ESP_GATTS_REG_EVT) {
-    if (param->reg.status == ESP_GATT_OK) {
-      gl_profile_tab[param->reg.app_id].gatts_if = gatts_if;
-    } else {
-      ESP_LOGI(GATTS_TAG, "Reg app failed, app_id %04x, status %d\n",
-               param->reg.app_id, param->reg.status);
-      return;
-    }
-  }
-
-  /* If the gatts_if equal to profile A, call profile A cb handler,
-   * so here call each profile's callback */
-  do {
-    int idx;
-    for (idx = 0; idx < PROFILE_NUM; idx++) {
-      if (gatts_if == ESP_GATT_IF_NONE || /* ESP_GATT_IF_NONE, not specify a
-                                             certain gatt_if, need to call every
-                                             profile cb function */
-          gatts_if == gl_profile_tab[idx].gatts_if) {
-        if (gl_profile_tab[idx].gatts_cb) {
-          gl_profile_tab[idx].gatts_cb(event, gatts_if, param);
-        }
-      }
-    }
-  } while (0);
-}
-
-gpio_num_t GPIO_INTR_A = static_cast<gpio_num_t>(16);
-gpio_num_t GPIO_INTR_B = static_cast<gpio_num_t>(5);
-#define GPIO_TAG "GPIO_ISR"
-
-static QueueHandle_t gpio_queue = NULL;
-
-static volatile int64_t last_interrupt_time = 0;
-
-#define DEBOUNCE_THRESHOLD_US 500000
-
-static void IRAM_ATTR gpio_isr_change_state_handler(void *arg) {
-  int gpio_num = (int)arg;
-
-  int64_t now = esp_timer_get_time();
-
-  if (now - last_interrupt_time > DEBOUNCE_THRESHOLD_US) {
-    last_interrupt_time = now;
-    xQueueSendFromISR(gpio_queue, &gpio_num, NULL);
-  }
-}
-
-static void gpio_isr_notify_change_state(void *arg) {
-  int io_num;
-
-  while (1) {
-    if (xQueueReceive(gpio_queue, &io_num, portMAX_DELAY)) {
-      ESP_LOGW(GPIO_TAG, "Interrupt on GPIO[%d]", io_num);
-
-      uint8_t value;
-      char *topic;
-
-      int profile = -1;
-
-      if (io_num == GPIO_INTR_A) {
-        profile = PROFILE_A_APP_ID;
-
-        if (state_A == 0) {
-          state_A = 1;
-        } else {
-          state_A = 0;
-        }
-        value = state_A;
-        topic = MQTT_TOPIC_A;
-      } else if (io_num == GPIO_INTR_B) {
-        profile = PROFILE_B_APP_ID;
-
-        if (state_B == 0) {
-          state_B = 1;
-        } else {
-          state_B = 0;
-        }
-        value = state_B;
-        topic = MQTT_TOPIC_B;
-      }
-
-      if (profile == -1) {
-        ESP_LOGE(GPIO_TAG, "Unknown GPIO in interrupt handler: %d", io_num);
-        return;
-      }
-
-      esp_ble_gatts_send_indicate(
-          gl_profile_tab[profile].gatts_if, gl_profile_tab[profile].conn_id,
-          gl_profile_tab[profile].char_handle, sizeof(value), &value, false);
-
-      char buffer[sizeof(uint8_t) * 8 + 1];
-      snprintf(buffer, sizeof(uint8_t) * 8 + 1, "%d", value);
-
-      // TODO send data with new mqtt module
-      // mqtt_send_data(topic, buffer);
-    }
-  }
-}
-
-extern "C" void start_gatt_server(void) {
-  esp_err_t ret;
-
-  ret = nvs_flash_init();
-  if (ret == ESP_ERR_NVS_NO_FREE_PAGES ||
-      ret == ESP_ERR_NVS_NEW_VERSION_FOUND) {
-    ESP_ERROR_CHECK(nvs_flash_erase());
-    ret = nvs_flash_init();
-  }
-  ESP_ERROR_CHECK(ret);
-
-  ESP_ERROR_CHECK(esp_bt_controller_mem_release(ESP_BT_MODE_CLASSIC_BT));
-
-  esp_bt_controller_config_t bt_cfg = BT_CONTROLLER_INIT_CONFIG_DEFAULT();
-  ret = esp_bt_controller_init(&bt_cfg);
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "%s initialize controller failed: %s\n", __func__,
-             esp_err_to_name(ret));
-    return;
-  }
-
-  ret = esp_bt_controller_enable(ESP_BT_MODE_BLE);
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "%s enable controller failed: %s\n", __func__,
-             esp_err_to_name(ret));
-    return;
-  }
-  ret = esp_bluedroid_init();
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "%s init bluetooth failed: %s\n", __func__,
-             esp_err_to_name(ret));
-    return;
-  }
-  ret = esp_bluedroid_enable();
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "%s enable bluetooth failed: %s\n", __func__,
-             esp_err_to_name(ret));
-    return;
-  }
-
-  ret = esp_ble_gatts_register_callback(gatts_event_handler);
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "gatts register error, error code = %x", ret);
-    return;
-  }
-  ret = esp_ble_gap_register_callback(gap_event_handler);
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "gap register error, error code = %x", ret);
-    return;
-  }
-  ret = esp_ble_gatts_app_register(PROFILE_A_APP_ID);
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "gatts app register error, error code = %x", ret);
-    return;
-  }
-  ret = esp_ble_gatts_app_register(PROFILE_B_APP_ID);
-  if (ret) {
-    ESP_LOGE(GATTS_TAG, "gatts app register error, error code = %x", ret);
-    return;
-  }
-  esp_err_t local_mtu_ret = esp_ble_gatt_set_local_mtu(500);
-  if (local_mtu_ret) {
-    ESP_LOGE(GATTS_TAG, "set local  MTU failed, error code = %x",
-             local_mtu_ret);
-  }
-
-  gpio_config_t io_conf_a = {
-      .pin_bit_mask = (1ULL << GPIO_INTR_A),
-      .mode = GPIO_MODE_INPUT,
-      .pull_up_en = GPIO_PULLUP_ENABLE,
-      .pull_down_en = GPIO_PULLDOWN_DISABLE,
-      .intr_type = GPIO_INTR_NEGEDGE,
-  };
-  gpio_config(&io_conf_a);
-
-  gpio_config_t io_conf_b = {
-      .pin_bit_mask = (1ULL << GPIO_INTR_B),
-      .mode = GPIO_MODE_INPUT,
-      .pull_up_en = GPIO_PULLUP_ENABLE,
-      .pull_down_en = GPIO_PULLDOWN_DISABLE,
-      .intr_type = GPIO_INTR_NEGEDGE,
-  };
-  gpio_config(&io_conf_b);
-
-  gpio_queue = xQueueCreate(10, sizeof(uint32_t));
-  if (gpio_queue == NULL) {
-    ESP_LOGE("GPIO", "Failed to create GPIO event queue");
-    return;
-  }
-
-  gpio_install_isr_service(0);
-
-  gpio_isr_handler_add(GPIO_INTR_A, gpio_isr_change_state_handler,
-                       (void *)GPIO_INTR_A);
-  gpio_isr_handler_add(GPIO_INTR_B, gpio_isr_change_state_handler,
-                       (void *)GPIO_INTR_B);
-
-  xTaskCreate(gpio_isr_notify_change_state, "gpio_isr_notify_change_state",
-              2048, NULL, 10, NULL);
-
-  return;
-}
