--- client/main/modules/gatt_server/gatt_server.h
@@ -1,14 +1,0 @@
-#ifndef GATT_SERVER_H
-#define GATT_SERVER_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void start_gatt_server(void);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif // GATT_SERVER_H
