--- client/main/modules/lcd/colors.h
@@ -1,150 +1,0 @@
-#define LCD_COLOR_ALICEBLUE 0xF7DF
-#define LCD_COLOR_ANTIQUEWHITE 0xFF5A
-#define LCD_COLOR_AQUA 0x07FF
-#define LCD_COLOR_AQUAMARINE 0x7FFA
-#define LCD_COLOR_AZURE 0xF7FF
-#define LCD_COLOR_BEIGE 0xF7BB
-#define LCD_COLOR_BISQUE 0xFF38
-#define LCD_COLOR_BLACK 0x0000
-#define LCD_COLOR_BLANCHEDALMOND 0xFF59
-#define LCD_COLOR_BLUE 0x001F
-#define LCD_COLOR_BLUEVIOLET 0x895C
-#define LCD_COLOR_BROWN 0xA145
-#define LCD_COLOR_BURLYWOOD 0xDDD0
-#define LCD_COLOR_CADETBLUE 0x5CF4
-#define LCD_COLOR_CHARTREUSE 0x7FE0
-#define LCD_COLOR_CHOCOLATE 0xD343
-#define LCD_COLOR_CORAL 0xFBEA
-#define LCD_COLOR_CORNFLOWERBLUE 0x64BD
-#define LCD_COLOR_CORNSILK 0xFFDB
-#define LCD_COLOR_CRIMSON 0xD8A7
-#define LCD_COLOR_CYAN 0x07FF
-#define LCD_COLOR_DARKBLUE 0x0011
-#define LCD_COLOR_DARKCYAN 0x03EF
-#define LCD_COLOR_DARKCYAN2 0x0451
-#define LCD_COLOR_DARKGOLDENROD 0xBC21
-#define LCD_COLOR_DARKGRAY 0xAD55
-#define LCD_COLOR_DARKGREEN2 0x0320
-#define LCD_COLOR_DARKGREEN 0x03E0
-#define LCD_COLOR_DARKGREY 0x7BEF
-#define LCD_COLOR_DARKKHAKI 0xBDAD
-#define LCD_COLOR_DARKMAGENTA 0x8811
-#define LCD_COLOR_DARKOLIVEGREEN 0x5345
-#define LCD_COLOR_DARKORANGE 0xFC60
-#define LCD_COLOR_DARKORCHID 0x9999
-#define LCD_COLOR_DARKRED 0x8800
-#define LCD_COLOR_DARKSALMON 0xECAF
-#define LCD_COLOR_DARKSEAGREEN 0x8DF1
-#define LCD_COLOR_DARKSLATEBLUE 0x49F1
-#define LCD_COLOR_DARKSLATEGRAY 0x2A69
-#define LCD_COLOR_DARKTURQUOISE 0x067A
-#define LCD_COLOR_DARKVIOLET 0x901A
-#define LCD_COLOR_DEEPPINK 0xF8B2
-#define LCD_COLOR_DEEPSKYBLUE 0x05FF
-#define LCD_COLOR_DIMGRAY 0x6B4D
-#define LCD_COLOR_DODGERBLUE 0x1C9F
-#define LCD_COLOR_FIREBRICK 0xB104
-#define LCD_COLOR_FLORALWHITE 0xFFDE
-#define LCD_COLOR_FORESTGREEN 0x2444
-#define LCD_COLOR_FUCHSIA 0xF81F
-#define LCD_COLOR_GAINSBORO 0xDEFB
-#define LCD_COLOR_GHOSTWHITE 0xFFDF
-#define LCD_COLOR_GOLD 0xFEA0
-#define LCD_COLOR_GOLDENROD 0xDD24
-#define LCD_COLOR_GRAY 0x8410
-#define LCD_COLOR_GREEN2 0x0400
-#define LCD_COLOR_GREEN 0x07E0
-#define LCD_COLOR_GREENYELLOW 0xAFE5
-#define LCD_COLOR_HONEYDEW 0xF7FE
-#define LCD_COLOR_HOTPINK 0xFB56
-#define LCD_COLOR_INDIANRED 0xCAEB
-#define LCD_COLOR_INDIGO 0x4810
-#define LCD_COLOR_IVORY 0xFFFE
-#define LCD_COLOR_KHAKI 0xF731
-#define LCD_COLOR_LAVENDER 0xE73F
-#define LCD_COLOR_LAVENDERBLUSH 0xFF9E
-#define LCD_COLOR_LAWNGREEN 0x7FE0
-#define LCD_COLOR_LEMONCHIFFON 0xFFD9
-#define LCD_COLOR_LIGHTBLUE 0xAEDC
-#define LCD_COLOR_LIGHTCORAL 0xF410
-#define LCD_COLOR_LIGHTCYAN 0xE7FF
-#define LCD_COLOR_LIGHTGOLDENRODYELLOW 0xFFDA
-#define LCD_COLOR_LIGHTGREEN 0x9772
-#define LCD_COLOR_LIGHTGREY 0xC618
-#define LCD_COLOR_LIGHTGREY2 0xD69A
-#define LCD_COLOR_LIGHTPINK 0xFDB8
-#define LCD_COLOR_LIGHTSALMON 0xFD0F
-#define LCD_COLOR_LIGHTSEAGREEN 0x2595
-#define LCD_COLOR_LIGHTSKYBLUE 0x867F
-#define LCD_COLOR_LIGHTSLATEGRAY 0x7453
-#define LCD_COLOR_LIGHTSTEELBLUE 0xB63B
-#define LCD_COLOR_LIGHTYELLOW 0xFFFC
-#define LCD_COLOR_LIME 0x07E0
-#define LCD_COLOR_LIMEGREEN 0x3666
-#define LCD_COLOR_LINEN 0xFF9C
-#define LCD_COLOR_MAGENTA 0xF81F
-#define LCD_COLOR_MAROON 0x7800
-#define LCD_COLOR_MAROON2 0x8000
-#define LCD_COLOR_MEDIUMAQUAMARINE 0x6675
-#define LCD_COLOR_MEDIUMBLUE 0x0019
-#define LCD_COLOR_MEDIUMORCHID 0xBABA
-#define LCD_COLOR_MEDIUMPURPLE 0x939B
-#define LCD_COLOR_MEDIUMSEAGREEN 0x3D8E
-#define LCD_COLOR_MEDIUMSLATEBLUE 0x7B5D
-#define LCD_COLOR_MEDIUMSPRINGGREEN 0x07D3
-#define LCD_COLOR_MEDIUMTURQUOISE 0x4E99
-#define LCD_COLOR_MEDIUMVIOLETRED 0xC0B0
-#define LCD_COLOR_MIDNIGHTBLUE 0x18CE
-#define LCD_COLOR_MINTCREAM 0xF7FF
-#define LCD_COLOR_MISTYROSE 0xFF3C
-#define LCD_COLOR_MOCCASIN 0xFF36
-#define LCD_COLOR_NAVAJOWHITE 0xFEF5
-#define LCD_COLOR_NAVY 0x000F
-#define LCD_COLOR_NAVY2 0x0010
-#define LCD_COLOR_OLDLACE 0xFFBC
-#define LCD_COLOR_OLIVE 0x7BE0
-#define LCD_COLOR_OLIVE2 0x8400
-#define LCD_COLOR_OLIVEDRAB 0x6C64
-#define LCD_COLOR_ORANGE 0xFD20
-#define LCD_COLOR_ORANGERED 0xFA20
-#define LCD_COLOR_ORCHID 0xDB9A
-#define LCD_COLOR_PALEGOLDENROD 0xEF55
-#define LCD_COLOR_PALEGREEN 0x9FD3
-#define LCD_COLOR_PALETURQUOISE 0xAF7D
-#define LCD_COLOR_PALEVIOLETRED 0xDB92
-#define LCD_COLOR_PAPAYAWHIP 0xFF7A
-#define LCD_COLOR_PEACHPUFF 0xFED7
-#define LCD_COLOR_PERU 0xCC27
-#define LCD_COLOR_PINK 0xF81F
-#define LCD_COLOR_PINK_2 0xFE19
-#define LCD_COLOR_PLUM 0xDD1B
-#define LCD_COLOR_POWDERBLUE 0xB71C
-#define LCD_COLOR_PURPLE 0x780F
-#define LCD_COLOR_PURPLE2 0x8010
-#define LCD_COLOR_RED 0xF800
-#define LCD_COLOR_ROSYBROWN 0xBC71
-#define LCD_COLOR_ROYALBLUE 0x435C
-#define LCD_COLOR_SADDLEBROWN 0x8A22
-#define LCD_COLOR_SALMON 0xFC0E
-#define LCD_COLOR_SANDYBROWN 0xF52C
-#define LCD_COLOR_SEAGREEN 0x2C4A
-#define LCD_COLOR_SEASHELL 0xFFBD
-#define LCD_COLOR_SIENNA 0xA285
-#define LCD_COLOR_SILVER 0xC618
-#define LCD_COLOR_SKYBLUE 0x867D
-#define LCD_COLOR_SLATEBLUE 0x6AD9
-#define LCD_COLOR_SLATEGRAY 0x7412
-#define LCD_COLOR_SNOW 0xFFDF
-#define LCD_COLOR_SPRINGGREEN 0x07EF
-#define LCD_COLOR_STEELBLUE 0x4416
-#define LCD_COLOR_TAN 0xD5B1
-#define LCD_COLOR_TEAL 0x0410
-#define LCD_COLOR_THISTLE 0xDDFB
-#define LCD_COLOR_TOMATO 0xFB08
-#define LCD_COLOR_TURQUOISE 0x471A
-#define LCD_COLOR_VIOLET 0xEC1D
-#define LCD_COLOR_WHEAT 0xF6F6
-#define LCD_COLOR_WHITE 0xFFFF
-#define LCD_COLOR_WHITESMOKE 0xF7BE
-#define LCD_COLOR_YELLOW 0xFFE0
-#define LCD_COLOR_YELLOWGREEN 0x9E66
