+++ server/.vscode/settings.json
@@ -81,6 +81,8 @@
         "semaphore": "cpp",
         "stop_token": "cpp",
         "thread": "cpp",
-        "cinttypes": "cpp"
+        "cinttypes": "cpp",
+        "complex": "cpp",
+        "netfwd": "cpp"
     },
 }
