--- server/partitions.csv
@@ -1,4 +1,0 @@
-# Name,   Type, SubType, Offset,  Size, Flags
-nvs,      data, nvs,     0x9000,  0x5000,
-phy_init, data, phy,     0xe000,  0x1000,
-factory,  app,  factory, 0x10000, 0x200000,
